Implementation

Repository Contents

The filesystem for the package code and documentation texts are organised in such a way as to be added in alphanumeric folder-then-contents order. This way, the complicated contents does not have to be managed by a plethora of include call managements for src/OceanSonar.jl, or the pages keyword for docs/make.jl.

OceanSonar Modelling Types

The field of ocean sonar consists of many metrics, each of which have many alternative contexts and methods for calculation. For simplicity of usage and the avoidance of namespace overpopulation, OceanSonar implements a collection of types that do much of the heavy lifting in the background.

Firstly:

OceanSonar.Model — Type

Model(name::AbstractString)
Model(name::Symbol)

Return Model{name}() which is used to dispatch functions by model names.

The name is stored as a Symbol in pascaltext style, i.e. only alphanumerals.

The implementation of Model is such that a basic user will never need to be aware of it. It is a specialised version of Base.Val defined specially for this package. Model cooperates with types defined in OceanSonar so users can call models by their name in a string or a symbol. For example:

julia> using OceanSonar
julia> import OceanSonar: sound_speed_profile
julia> sound_speed_profile(::Model{:MyModel}, z) = 1500 + exp(-z^2)(::OceanSonar.SoundSpeedProfileMetricFunctionType) (generic function with 7 methods)
julia> sound_speed_profile(:MyModel, 1e3)1500.0
julia> sound_speed_profile("My Model", 1e3)1500.0

Thus Model behaves as both a dispatching tool and a convenience enabler. For more advanced users, it is also accessible for easily extending MetricFunction instances.

To avoid piracy, this convenience behaviour is only defined for subtypes of OceanSonar.MetricFunction:

OceanSonar.MetricFunction — Type

OceanSonar.MetricFunction <: Function

Called instances dispatch its first argument as Symbol or AbstractString to Model.

E.g.

using OceanSonar
sound_speed isa OceanSonar.MetricFunction
import OceanSonar: sound_speed
sound_speed(::Model{:MyModel}, z::Real) = 1500.0 + exp(-z^2)
sound_speed(Model(:MyModel), 1e3)
sound_speed(:MyModel, 1e3)
sound_speed("My Model", 1e3)

See also: Model.

All metric functions defined by @implement_metric_function are listed in metric_functions:

OceanSonar.metric_functions — Constant

OceanSonar.metric_functions :: Vector{<:OceanSonar.MetricFunction}

Vector of MetricFunctions defined in OceanSonar.

using OceanSonar
OceanSonar.metric_functions

1-element Vector{OceanSonar.MetricFunction}:
 (::OceanSonar.SoundSpeedProfileMetricFunctionType) (generic function with 7 methods)

For each MetricFunction, you can view the Models implemented for it by running listmodels:

OceanSonar.listmodels — Function

listmodels(fcn::ModelFunction)

Returns the names of fcn models as a Vector of Strings formatted with titletext.

Each element of the output can be used as the first positional argument to fcn.

See also: listarguments, Model, OceanSonar.ModelFunction.

For each Model of a MetricFunction, you can view its inputs and parameters with listarguments:

OceanSonar.listarguments — Function

listarguments(fcn::ModelFunction, model::String)
listarguments(fcn::ModelFunction, model::Symbol)
listarguments(fcn::ModelFunction, model::Model)

Lists the inputs (positional arguments) and parameters (keyword arguments) of fcn for the model specified as a Vector of NamedTuples with fields inputs and parameters.

While InteractiveUtils.methodswith gives similar information, listarguments provides the argument names in a useable form.

Visualisation

The visual and visual! functions have methods described throughout the documentation, flexibly catering to the variety of types of plots that ocean sonar modelling benefits from.

Internally visual and visual! dispatch on the first argument's Model specification to specific Makie plotting recipes defined by the extension package OceanSonarMakieExt. Due to the nature of extensions, such plotting recipe functions are defined with no arguments in OceanSonar then imported for overloading in the extension package.

OceanSonar.visual — Function

visual(model, args...; kws...)
visual!(model, args...; kws...)
visual!(axis::Axis, model, args...; kws...)

Dispatches the args and kws to the appropriate plotting function, specified by model.

Has no plotting methods until Makie is loaded, which is normally done by loading one of its backends. Axis is from Makie.

The model can be a String, Symbol, or Model instance.

This wrapper function acts as both a convenience and avoidance for remembering the many plotting recipes defined in this package.

As per other OceanSonar modelling functions, a list of models is available via listmodels(visual) or listmodels(visual!)

OceanSonar.visual! — Function

visual(model, args...; kws...)
visual!(model, args...; kws...)
visual!(axis::Axis, model, args...; kws...)

Dispatches the args and kws to the appropriate plotting function, specified by model.

Has no plotting methods until Makie is loaded, which is normally done by loading one of its backends. Axis is from Makie.

The model can be a String, Symbol, or Model instance.

This wrapper function acts as both a convenience and avoidance for remembering the many plotting recipes defined in this package.

As per other OceanSonar modelling functions, a list of models is available via listmodels(visual) or listmodels(visual!)

using OceanSonar, CairoMakie
listmodels(visual)

1-element Vector{String}:
 "Sound Speed 2 D"

Supporting the creation of ocean metric visualisations is the OceanAxis2D function:

OceanSonar.OceanAxis2D — Function

OceanAxis2D

Behaves as Makie.Axis except with the default of yreversed = true.

Unusable until Makie is loaded.