Private Names

The following is an automated printing of documentation for OceanSonar.jl private names.

OceanSonar.KEEPTOKENS — Constant

OceanSonar.KEEPTOKENS

Used by textstyle. Vector of Strings of english language words (tokens) to keep tokenised and keep their text styles (for certain text style conversions by textstyle).

OceanSonar._sound_speed_profile_mediterranean_sea_linearised_itp — Constant

_sound_speed_profile_mediterranean_sea_linearised_itp Internal interpolation object.

OceanSonar._sound_speed_profile_north_atlantic_linearised_itp — Constant

_sound_speed_profile_north_atlantic_linearised_itp Internal interpolation object.

OceanSonar.metric_functions — Constant

OceanSonar.metric_functions :: Vector{<:OceanSonar.MetricFunction}

Vector of MetricFunctions defined in OceanSonar.

OceanSonar.textstyleseps — Constant

OceanSonar.textstyleseps

Used by textstyle. NamedTuple of separators (delimiters) for different text style types.

Core.String — Method

Overload of Core.Symbol for OceanSonar.Model types and instances.

Core.Symbol — Method

Overload of Core.Symbol for OceanSonar.Model types and instances.

OceanSonar.InstantiatingVisualFunctionType — Type

InstantiatingVisualFunctionType is the typeof(visual).

OceanSonar.MetricFunction — Type

OceanSonar.MetricFunction <: Function

Called instances dispatch its first argument as Symbol or AbstractString to Model.

E.g.

using OceanSonar
sound_speed isa OceanSonar.MetricFunction
import OceanSonar: sound_speed
sound_speed(::Model{:MyModel}, z::Real) = 1500.0 + exp(-z^2)
sound_speed(Model(:MyModel), 1e3)
sound_speed(:MyModel, 1e3)
sound_speed("My Model", 1e3)

See also: Model.

OceanSonar.ModelFunction — Type

ModelFunction

Abstract supertype of all OceanSonar's Model-dispatching functions.

OceanSonar.MutatingVisualFunctionType — Type

MutatingVisualFunctionType is the typeof(visual!).

OceanSonar._textstyle_casify_token — Method

_textstyle_casify_token(newstyle, token; keeptokens = OceanSonar.KEEPTOKENS)

Converts a single token to its specified newstyle.

OceanSonar._textstyle_verify_style — Method

_textstyle_verify_style

Used by textstyle. Verifies the request text style type.

OceanSonar.include_subroots — Method

OceanSonar.include_subroots(path::AbstractString)

Trade-off: If adding a file, need to restart REPL session.

OceanSonar.isalphanumeric — Method

OceanSonar.isalphanumeric(char::AbstractChar) -> Bool

Returns true if char is inclusively between:

'0' ≤ char ≤ '9'

'a' ≤ char ≤ 'z'

'A' ≤ char ≤ 'Z'

OceanSonar.soundspeedlines2d — Function

soundspeedlines2d
soundspeedlines2d!

Plotting functions. Methods to be populated by Makie extension.

OceanSonar.soundspeedlines2d! — Function

soundspeedlines2d
soundspeedlines2d!

Plotting functions. Methods to be populated by Makie extension.

OceanSonar.@implement_metric_function — Macro

@implement_metric_function function_name

Implements a new MetricFunction <: Function instance ready for Model dispatching.

Specifically:

Defines a concrete struct subtype of MetricFunction.

Defines a const instance of the new subtype with the name function_name.

Adds the new function to the list of metric_functions.

See also: Model, OceanSonar.MetricFunction.

Example

@implement_metric_function sound_speed_profile
sound_speed_profile(::Model{:MyModel}, z::Real) = 1500.0 + exp(-z^2)
sound_speed_profile("My Model", 1e3)