Text Case Styles

The function textstyle is an implementation of a rigorous heuristic for converting between text case styles.

Also provided are the convenience methods titletext, pascaltext, and snaketext.

OceanSonar.textstyle — Function

textstyle(
    newstyle :: Union{Symbol, <:AbstractString},
    oldtext :: Union{Symbol, <:AbstractString, <:Model{M}} where {M}
    ;
    keeptokens :: AbstractVector{<:AbstractString} = OceanSonar.keeptokens
)

Converts oldtext into the requested newstyle text.

Implemented text styles:

:space: Space-delimited tokens; tokens lowercased except keeptokens preserved.

:Title/:title/:Space: Space-delimited tokens; tokens' first character uppercased except keeptokens preserved.

:Pascal/:pascal: No token delimiter; tokens' first character uppercased except keeptokens preserved.

:camel: As :pascal but the very first character of text is lowercase.

:Snake: Underscore-delimited tokens; tokens' first character uppercased except keeptokens preserved.

:snake: Underscore-delimited tokens; tokens' first character lowercased except keeptokens preserved.

:Kebab: Hyphen-delimited tokens; tokens' first character uppercased except keeptokens.

:kebab: Hyphen-delimited tokens; tokens' first character lowercased except keeptokens.

Examples of implemented text styles:

:Space: "Say 32 Big Goodbyes to 1 Cruel NSW 1st World"

:space: "say 32 big goodbyes to 1 cruel NSW 1st world"

:pascal: "Say32BigGoodbyesTo1CruelNSW1stWorld"

:camel: "say32BigGoodbyesTo1CruelNSW1stWorld"

:Snake: "Say32BigGoodbyesto1CruelNSW1st_World"

:snake: "say32biggoodbyesto1cruelNSW1st_world"

:Kebab: "Say-32-Big-Goodbyes-to-1-Cruel-NSW-1st-World"

:kebab: "say-32-big-goodbyes-to-1-cruel-NSW-1st-world"

The title case behaves differently from Base.Unicode.titlecase, e.g.

julia> titlecase("say-32-big-goodbyes-to-1-cruel-NSW-1st-world")
"Say-32-Big-Goodbyes-To-1-Cruel-Nsw-1St-World"

julia> textstyle(:title, "say-32-big-goodbyes-to-1-cruel-NSW-1st-world")
"Say 32 Big Goodbyes to 1 Cruel NSW 1st World"

The following convenience methods are also exported:

titletext

snaketext

pascaltext

OceanSonar.titletext — Function

titletext(text) -> ::AbstractString
pascaltext(text) -> ::AbstractString
snaketext(text) -> ::AbstractString

Converts the inputted text to the named text case style.

Convenience functions for textstyle, internally calls e.g. textstyle(:title, text).

OceanSonar.pascaltext — Function

titletext(text) -> ::AbstractString
pascaltext(text) -> ::AbstractString
snaketext(text) -> ::AbstractString

Converts the inputted text to the named text case style.

Convenience functions for textstyle, internally calls e.g. textstyle(:title, text).

OceanSonar.snaketext — Function

titletext(text) -> ::AbstractString
pascaltext(text) -> ::AbstractString
snaketext(text) -> ::AbstractString

Converts the inputted text to the named text case style.

Convenience functions for textstyle, internally calls e.g. textstyle(:title, text).