Flexible Functions

At the time this documentation was written, NaNMath's functions were not designed for Intervals from IntervalArithmetic or Nums from Symbolics.

OceanSonar.Flex is designed to wrap functions of the same names in NaNMath and dispatch to the respectively appropriate "backend" - quoted because Nums and Intervals get passed to Base. "Normal" numbers passed to functions in NaNMath and OceanSonar.Flex will output NaN for invalid inputs e.g. sin(-1::Real) or sqrt(-1::Real), whereas the equivalent functions in Base throw an error.

OceanSonar.Flex — Module

OceanSonar.Flex :: Module

A module containing dispatching methods imported from other modules.

Specifically:

Contains mathematical functions that return NaN instead of erroring on mathematically invalid inputs,

such as sqrt(-1::Number).

derivative which dispatches appropriately

to DataInterpolations.derivative, ForwardDiff.derivative (not yet implemented), or Symbolics.derivative (not yet implemented).

The OceanSonar.Flexible functions are a temporary implementation until NaNMath functions are defined to fallback to their respective Base functions.

julia> sqrt(-1)ERROR: DomainError with -1.0:
sqrt was called with a negative real argument but will only return a complex result if called with a complex argument. Try sqrt(Complex(x)).

julia> sqrt(-1 |> complex)0.0 + 1.0im

julia> OceanSonar.Flex.sqrt(-1)NaN

julia> OceanSonar.Flex.sqrt(-1 |> complex)ERROR: MethodError: no method matching sqrt(::Complex{Int64})
You may have intended to import Base.sqrt
The function `sqrt` exists, but no method is defined for this combination of argument types.

Closest candidates are:
  sqrt(::ForwardDiff.Dual{T}) where T
   @ ForwardDiff ~/.julia/packages/ForwardDiff/PcZ48/src/dual.jl:238
  sqrt(::T) where T<:AbstractFloat
   @ NaNMath ~/.julia/packages/NaNMath/ceWIc/src/NaNMath.jl:17
  sqrt(::Real)
   @ NaNMath ~/.julia/packages/NaNMath/ceWIc/src/NaNMath.jl:18

julia> using OceanSonar, CairoMakie
julia> x = range(-1, 1, 51)-1.0:0.04:1.0
julia> lines(x, x .|> OceanSonar.Flex.sqrt)FigureAxisPlot()