OceanSonar.jl Usage

TODO: Link OceanSonar.jl definition names to their documentation.

Installation

As OceanSonar.jl is a Julia package, it requires the Julia programming language.

Install Julia using the instructions here.

Open a Julia REPL (terminal).

Activate the Pkg mode by pressing ] (at the beginning of the line).

Enter add https://github.com/kapple19/OceanSonar.jl which installs the OceanSonar.jl package.

Return to the Julian mode by backspacing (at the beginning of the line).

Enter using OceanSonar to load the package.

Registered packages require only add OceanSonar, which is not yet the case for this package under development.

Functions

The functions defined by OceanSonar.jl are:

atmosphere_celerity

ocean_celerity

seabed_celerity

shear_seabed_celerity

TODO: programmatically produce this list.

View their respective documentation for more details.

Containers

Data containers available for storing information are:

Medium

Slice

TODO: programmatically produce this list.

Functors

Functors are data containers that can be run like functions. The functor types defined by OceanSonar.jl available for instantiation are:

AtmosphereCelerity

OceanCelerity

SeabedCelerity

ShearSeabedCelerity

TODO: programmatically produce this list.

Their usage, for example:

using OceanSonar

models = list_models(OceanCelerity)
cel = OceanCelerity(models[1])

cel(0, 1e3)

Built-In Models

For any of the functions, containers, and functor types listed above, run list_models on them to list the available models for usage.

OceanSonar.list_models — Function

models::Symbol = list_models(
    func::Union{
        Function,
        Type{<:Container},
        Type{<:Result}
    }
)

Custom User Models

Models are implemented in Julia as methods signed with a Val{Symbol} as the first input. The Symbol instance must be snake-case. For example, to add custom models for ocean sound speed:

using OceanSonar
import OceanSonar: ocean_celerity

function ocean_celerity(::Val{:sinusoidal}, x, z)
    1500 + sin(x) - cos(z)
end

ocean_celerity(::Val{:depth_decreasing}, x, z) = 1500.0 + 20exp(-z)

which becomes available as

ocean_celerity("Sinusoidal", 1e2, 1e3)

and

ocean_celerity("Depth Decreasing", 0, 100)

Some models are interpolations of data points.

OceanSonar.Univariate — Type

Univariate is both an abstract type and a function that produces an interpolation-extrapolation object.

itp::Univariate = Univariate(args...; kwargs...)

Returns univariate interpolator-extrapolator object whose behaviour is specified by TODO.

TODO: Implement.

For example,

itp = Univariate(1:10, rand(10))
itp(2.4)

OceanSonar.Bivariate — Type

Bivariate <: Functor

Bivariate is both an abstract type and a function that produces an interpolation-extrapolation object.

itp::Bivariate = Univariate(args...; kwargs...)

Returns a bivariate interpolator-extrapolator object whose behaviour is specified by TODO.

TODO: Implement.

For example,

itp = Bivariate(1:10, 1:20, rand(10, 20))
itp(2.4, 11.5)

OceanSonar.Trivariate — Type

Trivariate is both an abstract type and a function that produces an interpolation-extrapolation object.

For example,

using OceanSonar
using CairoMakie

ranges = 1e3 * [0, 5, 6, 7, 10]
depths = 1e3 * [5, 5, 2, 5, 5]

itp = Univariate(ranges, depths)
bathymetry(::Val{:triangular_seamount}, x) = itp(x)

bty = Bathymetry("Triangular Seamount")
visual(bty)

Note that the pascal-case Bathymetry <: Univariate, but the Univariate instance is used to define the snake-case bathymetry method.

Visualisation

OceanSonar.jl comes with package extensions for backends Plots.jl and Makie.jl. The extensions define recipes for:

OceanCelerity

TODO: Programmatically produce this list.

So you can use either plotting backend's plotting functions on instances of the above-listed types.

OceanSonar.jl also provides visual which acts as a convenience function for plotting multiple features with one function call. Its mutating version visual! is also provided.

OceanSonar.visual! — Function

visual!(pos::Makie.GridPosition, args...)

Adds the plotting of args... to pos::Makie.GridPosition, where args... is the same as in visual.

See @doc visual.

OceanSonar.visual — Function

visual(args...)

Performs the following:

Creates a Makie.Figure with a single Makie.GridPosition.

Plots according to args...

Flips the y-axis positive-downwards for representing depth.

Plots the boundaries if such information is available (Slice and upward).

Adds a colorbar if 2D data (<:Bivariate) is plotted.

Adds labels and units.

Applies a default colouring for distinguishing between the different plots on the axis.

The first argument always specifies what the plotting target is. There are two ways to specify the plotting target: instances or types.

Plottable instances are:

Altimetry

Bathymetry

<:Celerity (e.g. OceanCelerity instances)

<:Density

Slice

Scenario

Beam

Vector{Beam}

Propagation

The second way to specify the plotting target is by arg[2] an instance of any of the above, and arg[1] an appropriate type contained by the instance inputted as arg[2].

Altimetry

Bathymetry

Boundary plots both the Altimetry and Bathymetry

OceanCelerity

OceanDensity

Beam

See documentation of the above-listed types for specific visual usage documentation.

Other Auxiliaries

Formatting a string of the model's name has been wrapped into the following functions:

OceanSonar.modelsnake — Function

modelsnake(model::Union{String, Symbol, Val})

Converts model as a String, Symbol, or Val into a snake-case String for OceanSonar model parsing.

OceanSonar.modeltitle — Function

modeltitle(model::Union{String, Symbol, Val})

Converts model as a String, Symbol, or Val into a title-case String for OceanSonar model parsing.

OceanSonar.modelsymbol — Function

modelsymbol(model::Union{String, Symbol, Val})

Converts model as a String, Symbol, or Val into a Symbol for OceanSonar model parsing.